requirements.txt
numpy
python-chess
torch==2.3.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
